check point dev

+++ src/controller/user-web-controller.tsx
@@ -2,12 +2,10 @@ import { Hono } from "hono";
 import {
   LoginUserRequest,
   RegisterUserRequest,
-  toUserResponse,
-  UpdateUserRequest,
 } from "../model/user-model";
 import { UserService } from "../service/user-service";
 import { ApplicationVariables } from "../model/app-model";
-import { renderer, LoginForm, RegisterForm } from "../view/layout-view";
+import { renderer, LoginForm, RegisterForm, UserPage } from "../view/layout-view";
 import { UserValidation } from "../validation/user-validation";
 import { prismaClient } from "../application/database";
 import { ZodError } from "zod";
@@ -15,36 +13,28 @@ import {
   ErrorMessageValidation,
   formatValidationErrors,
 } from "../validation/error-message-validation";
-import {
-  deleteCookie,
-  getCookie,
-  setCookie,
-  setSignedCookie,
-} from "hono/cookie";
-import { Session, User } from "@prisma/client";
+import { deleteCookie, getSignedCookie, setSignedCookie } from "hono/cookie";
+import { User } from "@prisma/client";
 import { webAuthMiddleware } from "../middleware/auth-middleware";
+
+
 export const userWebController = new Hono<{
   Variables: ApplicationVariables;
 }>();
 
-userWebController.get("*", renderer);
-const app_name = Bun.env.APP_NAME;
+const secret_session_key = Bun.env.SECRET_SESSION_KEY ?? "";
+
+userWebController.use("*", renderer);
 
 userWebController.get("/login", async (c) => {
   return c.render(
-    <div class="flex h-screen flex-col bg-red-500">
-      <h1 class="bg-blue-300">{app_name} Login</h1>
       <LoginForm email="" message={false} />
-    </div>
   );
 });
 
 userWebController.get("/register", async (c) => {
   return c.render(
-    <div class="flex h-screen flex-col bg-red-500">
-      <h1 class="bg-blue-300">{app_name} Register</h1>
       <RegisterForm email="" message={false} />
-    </div>
   );
 });
 
@@ -98,12 +88,12 @@ userWebController.post("/login", async (c) => {
       <LoginForm email={request.email} message={["cannot login session"]} />
     );
   }
-
-  setCookie(c, "x-hono-session", id, {
+  console.log("secret_session_key");
+  console.log(secret_session_key);
+  await setSignedCookie(c, "x-hono-session", id, secret_session_key, {
     path: "/",
     secure: true,
     httpOnly: true,
-    maxAge: 1000,
     expires: new Date(Date.now() + 6.048e8 * 1),
     sameSite: "Strict",
   });
@@ -160,11 +150,10 @@ userWebController.post("/register", async (c) => {
     );
   }
 
-  setCookie(c, "x-hono-session", id, {
+  await setSignedCookie(c, "x-hono-session", id, secret_session_key, {
     path: "/",
     secure: true,
     httpOnly: true,
-    maxAge: 1000,
     expires: new Date(Date.now() + 6.048e8 * 1),
     sameSite: "Strict",
   });
@@ -175,26 +164,28 @@ userWebController.post("/register", async (c) => {
 });
 
 userWebController.get("/logout", async (c) => {
-  const sessionId = getCookie(c, "x-hono-session");
-  deleteCookie(c, "x-hono-session", {
-    path: "/",
-    secure: true,
-  });
-  await UserService.deleteSession(sessionId);
+  const sessionId = await getSignedCookie(
+    c,
+    secret_session_key,
+    "x-hono-session"
+  );
+
+  if (!sessionId) {
+    return c.redirect("/login");
+  }
+
+  deleteCookie(c, "x-hono-session");
+  const result = await UserService.deleteSession(sessionId);
+
+  console.info("logout session : ", sessionId, ":", false);
 
   return c.redirect("/login");
 });
 
 userWebController.get("/user", webAuthMiddleware, async (c) => {
-  const user = c.get('user') as User;
+  const user = c.get("user") as User;
   return c.render(
-    <div class="flex h-screen flex-col bg-red-500">
-      <h1 class="bg-blue-300">{app_name} User Page</h1>
-      <p>{user.name ?? "-"}</p>
-      <p>{user.email ?? "-"}</p>
-      <p>{user.id ?? "-"}</p>
-      <p>{user.createdAt.toString() ?? "-"}</p>
-    </div>
+    <UserPage name={user.name} email={user.email} id={user.id} createdAt={user.createdAt} />
   );
 });
